requirements.txt
fastapi==0.114.0
uvicorn==0.34.0
line-bot-sdk==3.13.0
python-dotenv==1.0.1
pydantic==2.9.1
mysql-connector-python==9.3.0